+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this library are documented in this file.
 ### New Features
 
 - Add parsing support for CWSU Center Weather Advisory (#573).
+- Allow `pyiem.nws.nwsli` instance to be subscriptable for iterop.
 
 ### Bug Fixes